Update Source.cpp
Wrote main function, addItemToList function, and printInventory function. 
Need to fix error on line 24, code will not compile b/c setDiscount() is a non-static function and the scope resolution operator is trying to call it as if it were a static function. Looked at instructions and discountMultiplier has to be a static private member, and the main function has to be written as it is now.